+++ ser.cpp
@@ -6,14 +6,14 @@ namespace Internal
   {
     auto sz{static_cast<int32_t>(value.size())};
     strm.write(reinterpret_cast<const char *>(&sz), sizeof(sz));
-    strm.write(reinterpret_cast<const char *>(value.data()), sz);
+    strm.write(reinterpret_cast<const char *>(value.data()), static_cast<size_t>(sz));
   }
 
   auto deserVal(IStrm &strm, std::string &value) noexcept -> void
   {
     int32_t sz{};
     strm.read(reinterpret_cast<char *>(&sz), sizeof(sz));
-    value.resize(sz);
-    strm.read(reinterpret_cast<char *>(value.data()), sz);
+    value.resize(static_cast<size_t>(sz));
+    strm.read(reinterpret_cast<char *>(value.data()), static_cast<size_t>(sz));
   }
 } // namespace Internal
